+++ CHANGELOG.md
@@ -268,6 +268,8 @@
 - Added Monitor.count
 - Removed get_monitor_count
 - Added Monitor.new, Monitor[], and Monitor.all
+- Added Monitor.current
+- Removed get_current_monitor
 
 ## v0.3.14.1